Elimino definiciones innecesarias de Search para resolver conflictos
