+++ src/components/common/EventTabs.tsx
@@ -7,7 +7,13 @@ import { IEvent } from '@/data/domain/Event'
 import { Laptop } from '@phosphor-icons/react'
 import '../pages/search/search.css'
 
-function CustomTabPanel(props: { children?: React.ReactNode; value: number; index: number }) {
+interface TabPanelProps {
+  children?: React.ReactNode
+  index: number
+  value: number
+}
+
+function CustomTabPanel(props: TabPanelProps) {
   const { children, value, index, ...other } = props
 
   return (
@@ -21,6 +27,13 @@ function CustomTabPanel(props: { children?: React.ReactNode; value: number; inde
       {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
     </div>
   )
+}
+
+function a11yProps(index: number) {
+  return {
+    id: `simple-tab-${index}`,
+    'aria-controls': `simple-tabpanel-${index}`,
+  }
 }
 
 export default function EventTabs({ events }: { events: IEvent[] }) {
